requirements.txt
# ============================================================
# 1주차: HTTP 요청/응답 구조 확인
# ============================================================
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ============================================================
# 2주차: Pydantic 스키마 (fastapi 의존성으로 자동 설치됨 — 추가 패키지 없음)
# ============================================================

# ============================================================
# 3주차: DB 연동
# ============================================================
sqlalchemy>=2.0.0,<3
# aiosqlite==0.x         # 5주차: SQLite async 드라이버
# alembic==1.x           # DB 마이그레이션 (선택)

# ============================================================
# 4주차: LLM 연동
# ============================================================
langchain>=0.3.0,<1
langchain-openai>=0.2.0,<1
openai>=1.0.0,<2
python-dotenv>=1.0.0,<2

# ============================================================
# 5주차: 이미 async 구조가 fastapi + aiosqlite로 지원됨
# ============================================================
